% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{peg_sum}
\alias{peg_sum}
\title{Example PEG}
\usage{
peg_sum(x)
}
\arguments{
\item{x}{a vector of strings, each containing a comma-separated
set of digits}
}
\value{
a vector of numbers, containing either the sum of
the equivalent element of \code{x} or (if the element
could not be parsed) \code{NA}.
}
\description{
an example of a Parsing Expression Grammar
(PEG) that takes a comma-separated string of digits and
sums them together
}
\examples{
# Simple example
peg_sum("1,2, 5, 91, 34")

}
