% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piton.R
\docType{package}
\name{piton}
\alias{piton}
\alias{piton-package}
\title{Parsing Expression Grammars in Rcpp}
\description{
This package wraps the PEGTL library to make Parsing Expression Grammars
available to R/C++ developers. As an exported, header-only package, it can be included
in other Rcpp codebases using \code{depends} functionality, and is platform-independent.
}
\seealso{
the \href{https://github.com/Ironholds/piton/blob/master/README.md}{README},
or \code{\link{peg_sum}} for an example.
}
